code/src/requirements.txt
streamlit
pandas
scikit-learn
transformers
datasets
evaluate
torch
PyPDF2
google-generativeai
fpdf